Another run at style. The feature here is clearing out some ugliness in use of reflection and newInstance(). Fixed up some copyright headers.
git-svn-id: https://svn.apache.org/repos/asf/mahout/trunk@1196818 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/mahout/cf/taste/impl/recommender/GenericBooleanPrefUserBasedRecommender.java
@@ -54,7 +54,7 @@ public final class GenericBooleanPrefUserBasedRecommender extends GenericUserBas
       // See GenericItemBasedRecommender.doEstimatePreference() too
       if (userID != theUserID && dataModel.getPreferenceValue(userID, itemID) != null) {
         foundAPref = true;
-        totalSimilarity += similarity.userSimilarity(theUserID, userID);
+        totalSimilarity += (float) similarity.userSimilarity(theUserID, userID);
       }
     }
     return foundAPref ? totalSimilarity : Float.NaN;
